requirements.txt
mediapipe
opencv-python
scikit-learn
pandas
numpy
seaborn
flask
flask_cors
fastapi
uvicorn
python-multipart
pillow